scoring species tree + weighting dupp versus loss

+++ main/phylonet/tree/model/sti/STITreeCluster.java
@@ -286,8 +286,8 @@ public class STITreeCluster
 		//public STITreeCluster _cluster = null;
 		//public int _el_num = -1;
 		//public int _min_cost = -1;
-		public int _max_score = Integer.MIN_VALUE;
-		public int _c = 0;
+		public double _max_score = Integer.MIN_VALUE;
+		public double _c = 0;
 		public Vertex _min_lc = this._min_rc = null;
 		public Vertex _min_rc;
 		public List<Vertex> _subcl = null;
